Move GetWindowsZOrder to IWindowingPlatform
