+++ src/Android/Avalonia.Android/Stubs.cs
@@ -12,6 +12,8 @@ namespace Avalonia.Android
         public IWindowImpl CreateEmbeddableWindow() => throw new NotSupportedException();
 
         public ITrayIconImpl? CreateTrayIcon() => null;
+
+        public void GetWindowsZOrder(ReadOnlySpan<IWindowImpl> windows, Span<long> zOrder) => throw new NotSupportedException();
     }
 
     internal class PlatformIconLoaderStub : IPlatformIconLoader
