+++ src/Avalonia.Controls/Platform/IWindowingPlatform.cs
@@ -1,3 +1,4 @@
+using System;
 using Avalonia.Metadata;
 
 namespace Avalonia.Platform
@@ -12,5 +13,14 @@ namespace Avalonia.Platform
         IWindowImpl CreateEmbeddableWindow();
 
         ITrayIconImpl? CreateTrayIcon();
+
+        /// <summary>
+        /// Fills a span with numbers that represent the relative order of the windows in the z-order.
+        /// The topmost window should have the highest number.
+        /// Both the <paramref name="windows"/> and <paramref name="zOrder"/> lists are expected to be the same length.
+        /// </summary>
+        /// <param name="windows">A span of windows to get their z-order.</param>
+        /// <param name="zOrder">The span to be filled with the associated window z-order.</param>
+        void GetWindowsZOrder(ReadOnlySpan<IWindowImpl> windows, Span<long> zOrder);
     }
 }
