+++ src/Headless/Avalonia.Headless/AvaloniaHeadlessPlatform.cs
@@ -63,6 +63,14 @@ namespace Avalonia.Headless
             public IWindowImpl CreateEmbeddableWindow() => throw new PlatformNotSupportedException();
 
             public ITrayIconImpl? CreateTrayIcon() => null;
+
+            public void GetWindowsZOrder(ReadOnlySpan<IWindowImpl> windows, Span<long> zOrder)
+            {
+                for (var i = 0; i < windows.Length; ++i)
+                {
+                    zOrder[i] = (windows[i] as HeadlessWindowImpl)?.ZOrder ?? 0;
+                }
+            }
         }
         
         internal static void Initialize(AvaloniaHeadlessPlatformOptions opts)
